Add wip code to browser.ts
