--- source/inject-proxy.ts
@@ -1,6 +1,0 @@
-import * as path from 'path';
-import {readFileSync} from 'fs';
-
-const s = document.createElement('script');
-s.innerHTML = readFileSync(path.join(__dirname, 'ws-proxy.js'), 'utf8').toString();
-document.documentElement.querySelector('body')?.append(s);
